Added Oracle support. Jar locking on Windows fixed.
To use Postgres postgres= instead of db= is required.
Added Revenj.NET for Postgres and Oracle as targets.

Oracle has strange licensing schema. Instruct users to download it or look into ORACLE_HOME for it.

Class loader doesn't have a close() on Java6. Invoke that method via reflection if possible.