CLC improvements
SBT update

Allow direct public usage of DatabaseInfo and creating from sql script (at least in Postgres)
Keep track of previous_hash. Cope with versions which do not have it. Drop the support for old -NGS- schema naming

Support for docs-metadata flag as an indication to embed documentation in code.